Updated classes
