+++ ObjectPrinting/IPropertyPrintingConfig.cs
@@ -1,0 +1,7 @@
+﻿namespace ObjectPrinting
+{
+    public interface IPropertyPrintingConfig<TOwner>
+    {
+        PrintingConfig<TOwner> ParentConfig { get; }
+    }
+}
