+++ ObjectPrinting/PrintingConfig.cs
@@ -87,51 +87,5 @@ namespace ObjectPrinting
         {
             return this;
         }
-    }
-    
-    public class PropertyPrintingConfig<TOwner, TPropType> : IPropertyPrintingConfig<TOwner>
-    {
-        private PrintingConfig<TOwner> parentConfig;
-
-        public PropertyPrintingConfig(PrintingConfig<TOwner> parentConfig)
-        {
-            this.parentConfig = parentConfig;
-        }
-
-        public PrintingConfig<TOwner> Using(Func<TPropType, string> func)
-        {
-            return parentConfig;
-        }
-
-
-        PrintingConfig<TOwner> IPropertyPrintingConfig<TOwner>.ParentConfig => parentConfig;
-    }
-
-    public interface IPropertyPrintingConfig<TOwner>
-    {
-        PrintingConfig<TOwner> ParentConfig { get; }
-    }
-
-    public static class PrintingConfigExtensions
-    {
-        public static PrintingConfig<TOwner> Using<TOwner>(this PropertyPrintingConfig<TOwner, int> config,
-            CultureInfo culture)
-        {
-            return (config as IPropertyPrintingConfig<TOwner>).ParentConfig;
-        }
-
-        public static PrintingConfig<TOwner> Substring<TOwner>(this PropertyPrintingConfig<TOwner, string> config,
-            int start, int end)
-        {
-            return (config as IPropertyPrintingConfig<TOwner>).ParentConfig;
-        }
-    }
-
-    public static class PrintingObjectExtensions
-    {
-        public static PrintingConfig<TOwner> Printing<TOwner>(this TOwner obj)
-        {
-            return new PrintingConfig<TOwner>(obj);
-        }
-    }
+    }  
 }
