+++ ObjectPrinting/PropertyPrintingConfig.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace ObjectPrinting
+{
+    public class PropertyPrintingConfig<TOwner, TPropType> : IPropertyPrintingConfig<TOwner>
+    {
+        private PrintingConfig<TOwner> parentConfig;
+
+        public PropertyPrintingConfig(PrintingConfig<TOwner> parentConfig)
+        {
+            this.parentConfig = parentConfig;
+        }
+
+        public PrintingConfig<TOwner> Using(Func<TPropType, string> func)
+        {
+            return parentConfig;
+        }
+        PrintingConfig<TOwner> IPropertyPrintingConfig<TOwner>.ParentConfig => parentConfig;
+    }
+}
